Manual conflict resolution and code fixes to make GFXR functional

+++ framework/decode/dx12_consumer_base.h
@@ -71,7 +71,7 @@ class Dx12ConsumerBase
                                                     uint32_t                                            width,
                                                     uint32_t                                            height,
                                                     uint32_t                                            stride,
-                                                    uint32_t                                            usage,
+                                                    uint64_t                                            usage,
                                                     uint32_t                                            layers,
                                                     const std::vector<format::HardwareBufferPlaneInfo>& plane_info)
     {}
